rename validation statistics module
